Convert Outcome column to integer type

+++ Membangun_model/modelling_tuning.py
@@ -42,7 +42,7 @@ DATA_PATH = os.path.join(BASE_DIR, "cleaned_pima_diabetes_fuad.csv")
 df = pd.read_csv(DATA_PATH)
 
 X = df.drop("Outcome", axis=1)
-y = df["Outcome"]
+y = df["Outcome"].astype(int) 
 
 X_train, X_test, y_train, y_test = train_test_split(
     X, y, test_size=0.2, random_state=42, stratify=y
